extract runtime types
